fix: rename reduser to Header and action to Header
