+++ src/reducers/header.js
@@ -2,7 +2,7 @@ import { CHANGE_MENU_STATUS } from "../constants";
 
 const BASIC_STATE = true;
 
-export const headerMenuStatus = (state = BASIC_STATE, { type }) => {
+export const header = (state = BASIC_STATE, { type }) => {
   if (type === CHANGE_MENU_STATUS) {
     return !state;
   } else {
